requirements.txt
numpy
biopython
scikit-learn
torch 
torchvision 
torchaudio
wandb 
